requirements.txt
absl-py==2.0.0
asttokens==2.4.1
attrs==23.1.0
blinker==1.7.0
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cmake==3.27.7
colorama==0.4.6
contourpy==1.2.0
cryptography==41.0.5
cvzone==1.6.1
cycler==0.12.1
dlib==19.24.2
docopt==0.6.2
executing==2.0.1
face-recognition==1.3.0
face-recognition-models==0.3.0
firebase-admin==6.2.0
Flask==3.0.0
Flask-JWT-Extended==4.5.3
flatbuffers==23.5.26
fonttools==4.45.1
google-api-core==2.14.0
google-api-python-client==2.108.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.13.1
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpcio==1.59.3
grpcio-status==1.59.3
httplib2==0.22.0
icecream==2.1.3
idna==3.6
importlib-metadata==6.8.0
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.5
MarkupSafe==2.1.3
matplotlib==3.8.2
mediapipe==0.10.8
msgpack==1.0.7
numpy==1.26.2
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
packaging==23.2
Pillow==10.1.0
pipreqs==0.4.13
proto-plus==1.22.3
protobuf==3.20.3
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.17.2
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
requests==2.31.0
rsa==4.9
screeninfo==0.8.1
six==1.16.0
sounddevice==0.4.6
uritemplate==4.1.1
urllib3==2.1.0
Werkzeug==3.0.1
yarg==0.1.9
zipp==3.17.0
